backend/requirements_ai.txt
# AI Models Dependencies
# Core ML/AI libraries
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.35.2
sentencepiece>=0.2.1

# Image processing
Pillow>=10.1.0

# Audio processing
pydub>=0.25.1

# Numerical computing
numpy>=1.24.3

# Scientific computing
scipy>=1.8.0

# Machine learning utilities
scikit-learn>=1.3.2

# Logging and utilities
tqdm>=4.27.0
